+++ flutter-idea/src/io/flutter/propertyeditor/PropertyEditorViewFactory.java
@@ -5,17 +5,10 @@
  */
 package io.flutter.propertyeditor;
 
-import com.intellij.ide.plugins.IdeaPluginDescriptor;
-import com.intellij.ide.plugins.PluginManagerCore;
 import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.extensions.PluginId;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowFactory;
-import com.intellij.ui.components.JBLabel;
-import com.intellij.ui.content.Content;
-import com.intellij.ui.content.ContentManager;
-import com.intellij.util.ui.UIUtil;
 import io.flutter.FlutterUtils;
 import io.flutter.actions.RefreshToolWindowAction;
 import io.flutter.bazel.WorkspaceCache;
@@ -27,18 +20,18 @@ import io.flutter.run.daemon.DevToolsService;
 import io.flutter.sdk.FlutterSdk;
 import io.flutter.sdk.FlutterSdkVersion;
 import io.flutter.utils.AsyncUtils;
+import io.flutter.view.ViewUtils;
 import kotlin.coroutines.Continuation;
 import org.jetbrains.annotations.NotNull;
 
-import javax.swing.*;
-import java.awt.*;
 import java.util.List;
 import java.util.Optional;
 
-import static io.flutter.sdk.FlutterSdkVersion.MIN_SUPPORTS_PROPERTY_EDITOR;
-
 public class PropertyEditorViewFactory implements ToolWindowFactory {
   @NotNull private static String TOOL_WINDOW_ID = "Flutter Property Editor";
+
+  @NotNull
+  private final ViewUtils viewUtils = new ViewUtils();
 
   @Override
   public Object isApplicableAsync(@NotNull Project project, @NotNull Continuation<? super Boolean> $completion) {
@@ -54,7 +47,7 @@ public class PropertyEditorViewFactory implements ToolWindowFactory {
 
     DartPluginVersion dartPluginVersion = DartPlugin.getDartPluginVersion();
     if (!dartPluginVersion.supportsPropertyEditor()) {
-      presentLabel(toolWindow, "Flutter Property Editor requires a newer version of the Dart plugin.");
+      viewUtils.presentLabel(toolWindow, "Flutter Property Editor requires a newer version of the Dart plugin.");
       return;
     }
 
@@ -94,26 +87,4 @@ public class PropertyEditorViewFactory implements ToolWindowFactory {
 
     toolWindow.setTitleActions(List.of(new RefreshToolWindowAction(TOOL_WINDOW_ID)));
   }
-
-  protected void presentLabel(ToolWindow toolWindow, String text) {
-    final JBLabel label = new JBLabel(text, SwingConstants.CENTER);
-    label.setForeground(UIUtil.getLabelDisabledForeground());
-    replacePanelLabel(toolWindow, label);
-  }
-
-  private void replacePanelLabel(ToolWindow toolWindow, JComponent label) {
-    ApplicationManager.getApplication().invokeLater(() -> {
-      final ContentManager contentManager = toolWindow.getContentManager();
-      if (contentManager.isDisposed()) {
-        return;
-      }
-
-      final JPanel panel = new JPanel(new BorderLayout());
-      panel.add(label, BorderLayout.CENTER);
-      final Content content = contentManager.getFactory().createContent(panel, null, false);
-      contentManager.removeAllContents(true);
-      contentManager.addContent(content);
-    });
-  }
-
 }
