+++ flutter-idea/src/io/flutter/view/ViewUtils.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2025 The Chromium Authors. All rights reserved.
+ * Use of this source code is governed by a BSD-style license that can be
+ * found in the LICENSE file.
+ */
+package io.flutter.view;
+
+import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.ui.VerticalFlowLayout;
+import com.intellij.openapi.wm.ToolWindow;
+import com.intellij.ui.components.JBLabel;
+import com.intellij.ui.components.labels.LinkLabel;
+import com.intellij.ui.content.Content;
+import com.intellij.ui.content.ContentManager;
+import com.intellij.util.ui.JBUI;
+import com.intellij.util.ui.UIUtil;
+import io.flutter.utils.LabelInput;
+
+import javax.swing.*;
+import java.awt.*;
+import java.util.List;
+
+public class ViewUtils {
+  public void presentLabel(ToolWindow toolWindow, String text) {
+    final JBLabel label = new JBLabel(text, SwingConstants.CENTER);
+    label.setForeground(UIUtil.getLabelDisabledForeground());
+    replacePanelLabel(toolWindow, label);
+  }
+
+  public void presentClickableLabel(ToolWindow toolWindow, List<LabelInput> labels) {
+    final JPanel panel = new JPanel(new GridLayout(0, 1));
+
+    for (LabelInput input : labels) {
+      if (input.listener == null) {
+        final JLabel descriptionLabel = new JLabel("<html>" + input.text + "</html>");
+        descriptionLabel.setBorder(JBUI.Borders.empty(5));
+        descriptionLabel.setHorizontalAlignment(SwingConstants.CENTER);
+        panel.add(descriptionLabel, BorderLayout.NORTH);
+      }
+      else {
+        final LinkLabel<String> linkLabel = new LinkLabel<>("<html>" + input.text + "</html>", null);
+        linkLabel.setBorder(JBUI.Borders.empty(5));
+        linkLabel.setListener(input.listener, null);
+        linkLabel.setHorizontalAlignment(SwingConstants.CENTER);
+        panel.add(linkLabel, BorderLayout.SOUTH);
+      }
+    }
+
+    final JPanel center = new JPanel(new VerticalFlowLayout(VerticalFlowLayout.CENTER));
+    center.add(panel);
+    replacePanelLabel(toolWindow, center);
+  }
+
+  private void replacePanelLabel(ToolWindow toolWindow, JComponent label) {
+    ApplicationManager.getApplication().invokeLater(() -> {
+      final ContentManager contentManager = toolWindow.getContentManager();
+      if (contentManager.isDisposed()) {
+        return;
+      }
+
+      final JPanel panel = new JPanel(new BorderLayout());
+      panel.add(label, BorderLayout.CENTER);
+      final Content content = contentManager.getFactory().createContent(panel, null, false);
+      contentManager.removeAllContents(true);
+      contentManager.addContent(content);
+    });
+  }
+}
